Fix wrong behavior / Bug on longpulse (Switch Plugin)
The old behavior was, that if a gpio request/set is done while a
longpulse runs the longpulse get not deleted. So, although a pin is set
to 1 it will set to 0 if longpulse ends.
With this patch the unused Par3 in setSystemTimer can be set to TMR_KILL
(1) which removes this timer. In switch plugin this is performed before
the gpio request gets executed.