Extend ADC Plugin with multiplex and powersave
If none of the 2 GPIO pins is set it works as usual and the value get
written in "Analog1". If one or both GPIO pins are set the pin is set to
HIGH while reading A0. This way 2 analog sensors are possible. The
positive source for the sensor is 1st GPIO for Analog1 and 2nd GPIO for
Analog2. It's also possible to set only one of the both GPIO. In this
case it works like a powersave because the sensor only get voltage for
the short moment of measure.