Addition of ESP8285 targets (#187)
* Addition of ESP8285 targets

In addition to the normal ESP8266 targets there are now targets for
normal_8285, test_8285 and dev_8285. Each of these targets provides an
additional build flag with -DESP8285

* Added ESP8285 build targets with correct flash size

Correction of the build modification done before, now with correct flash
siez of 1M

* Provide different GPIO dropdown lists for 8266 and 8285

The ESP8266 and ESP8285 expose different sets of GPIOs. This new version
of the web server provides different dropdown lists depending on the
build flag -DESP8285 (which is set in the new targets for building
ESP8285 fw).