Web form endoding
Using the latest core, the default encoding scheme fails. Maybe a bug in the current core or a bug in the 2.3.0 which they fixed. I cant tell. 

According to 
https://stackoverflow.com/questions/4526273/what-does-enctype-multipart-form-data-mean
the encoding scheme in the webforms can be 
    application/x-www-form-urlencoded (the default)
    multipart/form-data
    text/plain
When reading through it, it appears to me multipart/form-data is okay. And it fixes the issue.
Just because someone may know better, I introduced a #define.