ESP8266 setOutputPower
added config option to set the ESP8266's Wifi Output Power. I had  done this for a very old version of ESPEasy before but at that time, messed up the patch, so here's a new attempt.
It may not be a big energy saver, but in my case, it is necessary to reduce this setting (down to 5dBm from 20) in order for my PIR detectors to work correctly. Seems they are sensitive to the 2.4GHz interference at close quarters somehow.