[WiFi] Reduce CPU power on wifi connection issues
See this extensive test list: https://github.com/letscontrolit/ESPEasy/issues/2008#issuecomment-437661851

This patch will stop executing the scheduler for 2 connection attempts starting the 4th attempt.
So on attempt 4, 5, 8, 9, 12, 13, .... etc. will it stop executing the scheduler.
This will in the end result in undesired behavior, like not executing rules or acting on triggers when wifi connection is lost.
But that's for later concerns.