[Dallas 1wire] Add error state value selector
The Dallas 1wire sensor can get into some error state.
To signal an error state, a NaN was ouput, but that may be hard to use in rules.
This selector allows to set a fixed value to signal an error state.

Also removed a lot of redundant code and make sure no data is shared among other active instances of the same plugin.