[ESPeasy p2p] Show remote node as port description on device page
The node delivering the data for a plugin is now shown on the device page in the "Port" column.

Also added a fix to only allow updates from 1 node to a remote node.
Previously it was possible to have several nodes writing to the same remote p2p shared plugin.
This could lead to very strange results.

This also means that a remote installed plugin must now be deleted and re-installed to allow receiving data from a remote host if the remote host unit id has changed.