[WiFi] Fix No reconnect when setup page was not completed
The WiFi setup page does set a flag to indicate the connect attempt was initiated from the setup. 
But the flag was only reset when the setup page reloaded..
If that did not happen, the node could retry forever to (re)connect to the AP if the channel was changed or the node cannot see the old BSSID anymore (e.g. out of reach when using repeater)
Also the first attempt after changing WiFi settings would also (initially) fail due to retrying with old BSSID/channel combination.